+++ applovin_max/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Versions
 
 ## x.x.x
+* Add support to classify waterfall objects. For more info, check out our [docs](https://dash.applovin.com/documentation/mediation/flutter/getting-started/advanced-settings#waterfall-information-api).
 * Refactored error handling logic.
 ## 3.7.0
 * Depends on Android SDK v12.2.0 and iOS SDK v12.2.1.